project structure refactored
