+++ src/models/opencv_face_detector.pbtxt
